Replace the Watcher class with python-essentials Watcher
This is an almost exact copy of the Watcher class from the
python-essentials repo (with small function/attribute name changes to
match the general 'responder' instead of 'file uploader' functionality.

I suspect the original version of this script was copied from the old
version of the python-essentials sync (before a refactor). The newer
version has some added functionality in that it also notices when new
files are added.