Further sync script refactors (#3)
* Replace the Watcher class with python-essentials Watcher

This is an almost exact copy of the Watcher class from the
python-essentials repo (with small function/attribute name changes to
match the general 'responder' instead of 'file uploader' functionality.

I suspect the original version of this script was copied from the old
version of the python-essentials sync (before a refactor). The newer
version has some added functionality in that it also notices when new
files are added.

* Replace the FileUploader class with python-essentials equivalent

Like the previous commit, this uses the latest version of the
FileUploader from the python-essentials repo (in that repo it's called
Updater but serves the exact same purpose). This is cleaner / flatter
and I think represents better code that we'd be happier for users to
see.