Make postgres connection lazy
Previously this caused problems because database access was attempted
before the database was up, causing the service to fail. This refactor
doesn't prevent that issue, but it at least means that connection won't
be attempted until needed. More importantly, it separates out the
connection into a separate function and makes error handling easier