Make postgres service more robust (#5)
* Update PgExec to take connection string as an argument
* Make postgres connection lazy

Previously this caused problems because database access was attempted
before the database was up, causing the service to fail. This refactor
doesn't prevent that issue, but it at least means that connection won't
be attempted until needed. More importantly, it separates out the
connection into a separate function and makes error handling easier

* Add retry to the database connection